requirements.txt
python-docx==1.1.2
flask==2.0.1
gunicorn==20.1.0 